Preserve terra server set as duplicate of terra config set server.
